[ADEV-82] Update campaign detail page design
