+++ frontend/src/css/ionic-override.css
@@ -270,5 +270,6 @@ ion-progress-bar::part(progress)::after {
   display: block;
   position: absolute;
   top: 0;
-  left: 100%;
+  right: 0;
+  transform: translateX(100%);
 }
