--- frontend/src/view/annotation-campaign-detail.page.css
@@ -1,4 +1,0 @@
-.archive-description {
-  text-align: center;
-  color: var(--ion-color-medium);
-}
