+++ frontend/src/view/detail-campaign/annotation-campaign-detail.page.css
@@ -1,0 +1,113 @@
+#detail-campaign {
+  display: grid;
+  gap: 4rem;
+  justify-content: center;
+  justify-items: stretch;
+  font-family: var(--ion-font-family);
+}
+
+
+#detail-campaign > #head {
+  display: grid;
+  justify-items: center;
+}
+
+#detail-campaign > #head h2 {
+  margin: 0;
+}
+
+#detail-campaign > #head .archive-description {
+  color: var(--ion-color-medium);
+}
+
+
+#detail-campaign > .bloc {
+  display: grid;
+  border: 1px solid rgba(var(--ion-color-medium-rgb), 0.5);
+  padding: 1rem;
+  gap: 1rem;
+  border-radius: 1rem;
+}
+
+#detail-campaign > .bloc .head-bloc {
+  display: grid;
+  grid-auto-flow: column;
+  justify-content: space-between;
+  align-items: center;
+  gap: 1rem;
+}
+#detail-campaign > .bloc .table-bloc {
+  text-align: center;
+  display: grid;
+  grid-template-columns: 1fr minmax(auto, 192px);
+  gap: 0.5rem 0;
+}
+#detail-campaign > .bloc .table-bloc .table-bloc-head {
+  font-weight: 300;
+}
+#detail-campaign > .bloc .table-bloc .table-bloc-head,
+#detail-campaign > .bloc .table-bloc .table-bloc-content {
+  padding: 0.25rem 1rem;
+  display: grid;
+  align-content: center;
+}
+#detail-campaign > .bloc .table-bloc .table-bloc-content:nth-child(3n + 2) {
+  justify-items: start;
+}
+#detail-campaign > .bloc .table-bloc .table-bloc-head:nth-child(2n + 1),
+#detail-campaign > .bloc .table-bloc .table-bloc-content:nth-child(3n + 1) {
+  border-right: 1px solid rgba(var(--ion-color-medium-rgb), 0.5);
+}
+#detail-campaign > .bloc .table-bloc .divider {
+  height: 1px;
+  background: rgba(var(--ion-color-medium-rgb), 0.5);
+  grid-column: 1 / 3;
+}
+
+#detail-campaign h5 {
+  margin: 0;
+}
+
+
+#detail-campaign > #global-info {
+  grid-template-columns: 1fr 1fr;
+  justify-items: start;
+}
+
+#detail-campaign > #global-info h5 {
+  justify-self: center;
+  grid-column: 1 / 3;
+}
+
+
+#detail-campaign > #global-info .item {
+  display: grid;
+  gap: 0.5rem;
+  grid-auto-flow: column;
+  align-items: center;
+  font-weight: 500;
+}
+
+#detail-campaign > #global-info .item .icon {
+  opacity: 0.5;
+}
+
+#detail-campaign > #global-info .item .label,
+#detail-campaign > #global-info .description .label {
+  font-weight: 300;
+}
+
+#detail-campaign > #global-info .description {
+  grid-column: 1 / 3;
+}
+
+#detail-campaign > #global-info .buttons {
+  grid-column: 1 / 3;
+}
+
+#detail-campaign .buttons {
+  justify-self: end;
+  display: grid;
+  grid-auto-flow: column;
+  gap: 1rem;
+}
